Zephyr module definition make file and c wrapper
Make file that builds the module as a lib using the zephyr make
system. It enforces use of C++11 as that is required for the
library.
As the library is c++ a wrapper is needed to be able to call it
from the Zephyr C applications. This change includes a wrapper for
the common usage of the lib in a Zephyr context but can be extended
as needed. The wrappers do not have a copy of the API documentation
but refers to the c++ documentation with a few comments relevant
for the usage.
This commit is tested using the posix build of the LE-Audio unicast
test applications with encoded LC3 data transferred over a Bleutooth
CIS link. As a change ID is needed in this repo to be able to build
the Zephyr code this change needs to be merged before the zephyr
change can build.

Signed-off-by: Casper Bonde <casper_bonde@bose.com>